+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.0 (2017-09-11)
+- Update rollup to 0.49.
+- __Breaking:__ Changes in options, see https://rollupjs.org/#big-list-of-options
+- Add rollup to peerDependencies.
+
 ## 0.3.0 (2017-08-01)
 - __Breaking:__ Force named exports even with a single default export.
 - Add installation and usage guide in the `README.md`.